minor fixes (new bug on cards)
