+++ components/layout/Layout.js
@@ -11,16 +11,12 @@ export default function Layout({ children, title = null }) {
   return (
     <>
       <Head>
-        <title>{title ? `${title} - Next Starter` : 'Next Starter'}</title>
+        <title>{title ? `SH.Q.K Forina - ${title}` : 'SH.Q.K Forina'}</title>
 
         {/* Favicons */}
-        <link rel="icon" href="/favicons/fav-32.ico" sizes="32x32" />
-        <link rel="icon" href="/favicons/fav-57.ico" sizes="57x57" />
-        <link rel="icon" href="/favicons/fav-76.ico" sizes="76x76" />
-        <link rel="icon" href="/favicons/fav-96.ico" sizes="96x96" />
-        <link rel="icon" href="/favicons/fav-128.ico" sizes="128x128" />
-        <link rel="icon" href="/favicons/fav-192.ico" sizes="192x192" />
-        <link rel="icon" href="/favicons/fav-228.ico" sizes="228x228" />
+        <link rel="icon" href="/favicons/favicon-32x-32" sizes="32x32" />
+        <link rel="icon" href="/favicons/android-chrome-192x192.png" sizes="192x192" />
+        <link rel="icon" href="/favicons/android-chrome-256x256.png" sizes="228x228" />
       </Head>
       <div className={`font-sans antialiased flex flex-col ${dev ? 'debug-screens' : ''}`}>
         <Navigation />
