--- content/cardItems.js
@@ -1,17 +1,0 @@
-export const cardItems = [
-  {
-    src: '/svg/hat-image.svg',
-    text: 'Mësim Bashkëkohor',
-    className: 'bg-blue-950',
-  },
-  {
-    src: '/svg/book-image.svg',
-    text: 'Bibliotekë',
-    className: 'bg-red-950',
-  },
-  {
-    src: '/svg/ball-image.svg',
-    text: 'Fushë Sporti',
-    className: 'bg-orange-950',
-  },
-]
