--- content/gallery.js
@@ -1,116 +1,0 @@
-export const gallery = [
-  {
-    src: '/img/1.jpeg',
-    width: 500,
-    height: 500,
-    alt: 'GalleryImages1',
-  },
-  {
-    src: '/img/2.jpeg',
-    width: 500,
-    height: 500,
-    alt: 'GalleryImages2',
-  },
-  {
-    src: '/img/3.jpeg',
-    width: 500,
-    height: 500,
-    alt: 'GalleryImages3',
-  },
-  {
-    src: '/img/4.jpeg',
-    width: 500,
-    height: 500,
-    alt: 'GalleryImages4',
-  },
-  {
-    src: '/img/5.jpeg',
-    width: 500,
-    height: 500,
-    alt: 'GalleryImages5',
-  },
-  {
-    src: '/img/6.jpeg',
-    width: 500,
-    height: 500,
-    alt: 'GalleryImages6',
-  },
-  {
-    src: '/img/7.jpeg',
-    width: 500,
-    height: 500,
-    alt: 'GalleryImages7',
-  },
-  {
-    src: '/img/8.jpeg',
-    width: 500,
-    height: 500,
-    alt: 'GalleryImages8',
-  },
-  {
-    src: '/img/9.jpeg',
-    width: 500,
-    height: 500,
-    alt: 'GalleryImages9',
-  },
-  {
-    src: '/img/10.jpeg',
-    width: 500,
-    height: 500,
-    alt: 'GalleryImages10',
-  },
-  {
-    src: '/img/11.jpeg',
-    width: 500,
-    height: 500,
-    alt: 'GalleryImages11',
-  },
-  {
-    src: '/img/12.jpeg',
-    width: 500,
-    height: 500,
-    alt: 'GalleryImages12',
-  },
-  {
-    src: '/img/13.jpg',
-    width: 500,
-    height: 500,
-    alt: 'GalleryImages12',
-  },
-  {
-    src: '/img/14.jpg',
-    width: 500,
-    height: 500,
-    alt: 'GalleryImages12',
-  },
-  {
-    src: '/img/15.jpg',
-    width: 500,
-    height: 500,
-    alt: 'GalleryImages12',
-  },
-  {
-    src: '/img/16.jpg',
-    width: 500,
-    height: 500,
-    alt: 'GalleryImages12',
-  },
-  {
-    src: '/img/17.jpg',
-    width: 500,
-    height: 500,
-    alt: 'GalleryImages12',
-  },
-  {
-    src: '/img/18.jpg',
-    width: 500,
-    height: 500,
-    alt: 'GalleryImages12',
-  },
-  {
-    src: '/img/19.jpg',
-    width: 500,
-    height: 500,
-    alt: 'GalleryImages12',
-  },
-]
