+++ content/home.js
@@ -1,0 +1,199 @@
+export const cardData = [
+  {
+    src: '/svg/hat-image.svg',
+    text: 'Mësim Bashkëkohor',
+    className: 'bg-blue-950',
+  },
+  {
+    src: '/svg/book-image.svg',
+    text: 'Bibliotekë',
+    className: 'bg-red-950',
+  },
+  {
+    src: '/svg/ball-image.svg',
+    text: 'Fushë Sporti',
+    className: 'bg-blue-400',
+  },
+]
+
+export const gallery = [
+  {
+    src: '/img/1.jpeg',
+    width: 500,
+    height: 500,
+    alt: 'GalleryImages1',
+  },
+  {
+    src: '/img/2.jpeg',
+    width: 500,
+    height: 500,
+    alt: 'GalleryImages2',
+  },
+  {
+    src: '/img/3.jpeg',
+    width: 500,
+    height: 500,
+    alt: 'GalleryImages3',
+  },
+  {
+    src: '/img/4.jpeg',
+    width: 500,
+    height: 500,
+    alt: 'GalleryImages4',
+  },
+  {
+    src: '/img/5.jpeg',
+    width: 500,
+    height: 500,
+    alt: 'GalleryImages5',
+  },
+  {
+    src: '/img/6.jpeg',
+    width: 500,
+    height: 500,
+    alt: 'GalleryImages6',
+  },
+  {
+    src: '/img/7.jpeg',
+    width: 500,
+    height: 500,
+    alt: 'GalleryImages7',
+  },
+  {
+    src: '/img/8.jpeg',
+    width: 500,
+    height: 500,
+    alt: 'GalleryImages8',
+  },
+  {
+    src: '/img/9.jpeg',
+    width: 500,
+    height: 500,
+    alt: 'GalleryImages9',
+  },
+  {
+    src: '/img/10.jpeg',
+    width: 500,
+    height: 500,
+    alt: 'GalleryImages10',
+  },
+  {
+    src: '/img/11.jpeg',
+    width: 500,
+    height: 500,
+    alt: 'GalleryImages11',
+  },
+  {
+    src: '/img/12.jpeg',
+    width: 500,
+    height: 500,
+    alt: 'GalleryImages12',
+  },
+  {
+    src: '/img/13.jpg',
+    width: 500,
+    height: 500,
+    alt: 'GalleryImages12',
+  },
+  {
+    src: '/img/14.jpg',
+    width: 500,
+    height: 500,
+    alt: 'GalleryImages12',
+  },
+  {
+    src: '/img/15.jpg',
+    width: 500,
+    height: 500,
+    alt: 'GalleryImages12',
+  },
+  {
+    src: '/img/16.jpg',
+    width: 500,
+    height: 500,
+    alt: 'GalleryImages12',
+  },
+  {
+    src: '/img/17.jpg',
+    width: 500,
+    height: 500,
+    alt: 'GalleryImages12',
+  },
+  {
+    src: '/img/18.jpg',
+    width: 500,
+    height: 500,
+    alt: 'GalleryImages12',
+  },
+  {
+    src: '/img/19.jpg',
+    width: 500,
+    height: 500,
+    alt: 'GalleryImages12',
+  },
+]
+
+export const homeImages = [
+  {
+    src: '/img/1.jpeg',
+    width: '500',
+    height: '500',
+    alt: 'ShkollaForinë',
+  },
+  {
+    src: '/img/2.jpeg',
+    width: '500',
+    height: '500',
+    alt: 'ShkollaForinë1',
+  },
+  {
+    src: '/img/3.jpeg',
+    width: '500',
+    height: '500',
+    alt: 'ShkollaForinë2',
+  },
+  {
+    src: '/img/4.jpeg',
+    width: '500',
+    height: '500',
+    alt: 'ShkollaForinë3',
+  },
+]
+
+export const programdata = [
+  {
+    text: 'Plan Programi Vjetor',
+    href: '/documents/Planprogrami-Vjetor-2023-2024-Forinë-SHQIP.doc',
+    download: '/documents/Planprogrami-Vjetor-2023-2024-Forinë-SHQIP.doc',
+  },
+  {
+    text: 'Orari i Punës - Para dite',
+    href: '/documents/Orari-i-Punës-Para-Dite-23-24.docx',
+    download: '/documents/Orari-i-Punës-Para-Dite-23-24.docx',
+  },
+  {
+    text: 'Orari i Punës - Pas dite',
+    href: '/documents/Orari-i-Punës-Pas-Dite-23-24.docx',
+    download: '/documents/Orari-i-Punës-Pas-Dite-23-24.docx',
+  },
+  {
+    text: 'Plani Zhvillimor',
+    href: '/documents/Plani-Zhvillimor.docx',
+    download: '/documents/Plani-Zhvillimor.docx',
+  },
+  {
+    text: 'Raport Vjetor 2022-2023',
+    href: '/documents/Raport-Vjetor-2022-2023.docx',
+    download: '/documents/Raport-Vjetor-2022-2023.docx',
+  },
+  {
+    text: 'Годишна програма',
+    href: '/documents/Годишна-програма-на-основното-училиште-2023-2024.docx',
+    download: '/documents/Годишна-програма-на-основното-училиште-2023-2024.docx',
+  },
+  {
+    text: 'Programa Foto',
+    href: '/documents/20231002_113650.jpg',
+    download: '/documents/20231002_113650.jpg',
+  },
+]
