--- content/homeImages.js
@@ -1,26 +1,0 @@
-export const homeImages = [
-  {
-    src: '/img/1.jpeg',
-    width: '500',
-    height: '500',
-    alt: 'ShkollaForinë',
-  },
-  {
-    src: '/img/2.jpeg',
-    width: '500',
-    height: '500',
-    alt: 'ShkollaForinë1',
-  },
-  {
-    src: '/img/3.jpeg',
-    width: '500',
-    height: '500',
-    alt: 'ShkollaForinë2',
-  },
-  {
-    src: '/img/4.jpeg',
-    width: '500',
-    height: '500',
-    alt: 'ShkollaForinë3',
-  },
-]
