+++ package.json
@@ -15,7 +15,8 @@
     "autoprefixer": "^10.4.13",
     "next": "^13.2.1",
     "next-seo": "^5.15.0",
-    "postcss": "8.4.21",
+    "node": "^20.7.0",
+    "postcss": "^8.4.31",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "tailwindcss": "^3.2.7"
