+++ public/next-sitemap.js
@@ -1,0 +1,7 @@
+/* eslint-disable prettier/prettier */
+module.exports = {
+  siteUrl: process.env.SITE_URL || 'https://www.shfkforina.edu.mk/',
+  generateRobotsTxt: false,
+  sitemapSize: 7000,
+  changefreq: 'weekly',
+}
