+++ tailwind.config.js
@@ -25,16 +25,11 @@ module.exports = {
         yellow: {
           ...config.colors.yellow,
           950: '#E4A40F',
-          960: '#FAF8D4',
+          1000: '#FAF8D4',
         },
-
         blue: {
           ...config.colors.blue,
           950: '#035392',
-        },
-        orange: {
-          ...config.colors.orange,
-          950: '#FF971D',
         },
       },
     },
